検索結果のリストを切り出す
